examples: move json server and client to advanced package
Fixes: #1778 (again)
